Post entry
I added the possibility to add an entry to the agenda. I also fixed the NewSubstance adapter to work with a more accessible class.

I also moved the email validity check to the press of the login/register button. That way users won't get a negative feeling when entering their email. Which we found out, was happening.